+++ src/main/java/org/jenkinsci/plugins/DependencyTrack/JobAction.java
@@ -58,6 +58,7 @@ public class JobAction extends InvisibleAction {
      */
     @JavaScriptMethod
     public JSONArray getSeverityDistributionTrend() {
+        project.checkPermission(Job.READ);
         final List<SeverityDistribution> severityDistributions = project.getBuilds().stream()
                 .sorted(Comparator.naturalOrder())
                 .map(run -> run.getAction(ResultAction.class)).filter(Objects::nonNull)