+++ src/main/java/org/jenkinsci/plugins/DependencyTrack/ResultAction.java
@@ -19,6 +19,7 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import hudson.Plugin;
 import hudson.PluginWrapper;
 import hudson.model.Action;
+import hudson.model.Job;
 import hudson.model.Run;
 import java.io.Serializable;
 import java.util.Collection;
@@ -110,6 +111,7 @@ public class ResultAction implements RunAction2, SimpleBuildStep.LastBuildAction
      */
     @JavaScriptMethod
     public JSONArray getFindingsJson() {
+        run.checkPermission(Job.READ);
         return JSONArray.fromObject(findings);
     }